Fork: add Salling Group style
Now that existing style configuration is fully generified, define a new
"salling-group" style; this style is the purpose of our fork.

The Salling Group style currently looks more like conventional Java
formatting than Google's two styles.

I would have preferred to capture our legacy single-tab formatting and
show evolution but this requires extending sg-java-format with more
complex indentation control and I'm not willing to do that.

The 80 column limit is an experiment. It's my personal preference, by
observation that manual adherence produces entirely legible code, but
it's controversial to be so conservative and it is possible that a limit
of 80 works poorly with the rest of google-java-format. If this limit is
found to be unsustainable, try 85-90 instead of going straight to 100.