Fork: rename JAR file
To avoid confusion and name clashes between google-java-format and our
fork, rename the generated JAR file. Use just an "sg" prefix instead of
our full company name, which is rather unwieldy.

There are numerous ways to achieve this, with various problems and
resolutions [1, 2, 3]. This commit changes the actual group ID and
artifact name -- it's not strictly necessary but it's most honest and
avoids generating a "fake" google-java-format binary.

[1] https://stackoverflow.com/questions/4238944/controlling-maven-final-name-of-jar-artifact
[2] https://maven.apache.org/plugins/maven-shade-plugin/shade-mojo.html
[3] https://stackoverflow.com/questions/4088701/maven-shaded-jar-is-prefixed-with-original-in-the-file-name